refactor jwt middleware to base.
[changelog]
internal: refactor jwt middleware for code reusability.
